requirements.txt
click==8.1.7
colorama==0.4.6
PyYAML==6.0.2
